refactor: move posts service into contentlayer lib
