+++ src/pages/blog/index.tsx
@@ -6,10 +6,10 @@ import { RiSearch2Line } from 'react-icons/ri'
 import PageSeo from '@/components/PageSeo'
 import PageTitle from '@/components/PageTitle'
 import { PostCard } from '@/components/PostCard'
-import { filterPosts, getAllPosts } from '@/services/posts'
+import { filterPosts, getPosts } from '@/lib/contentlayer'
 
 export const getStaticProps = async () => {
-  const allPosts = getAllPosts([
+  const allPosts = getPosts([
     'title',
     'date',
     'slug',
