+++ src/pages/tags/[tag].tsx
@@ -1,5 +1,3 @@
-// import { getAllPosts } from '@/services/posts'
-
 import { motion } from 'framer-motion'
 import type {
   GetStaticPaths,
@@ -12,8 +10,8 @@ import { RiSearch2Line } from 'react-icons/ri'
 import PageSeo from '@/components/PageSeo'
 import PageTitle from '@/components/PageTitle'
 import { PostCard } from '@/components/PostCard'
+import { filterPosts, getPosts } from '@/lib/contentlayer'
 import { kebabCase } from '@/lib/utils'
-import { filterPosts, getAllPosts } from '@/services/posts'
 import { getAllTags } from '@/services/tags'
 
 export const getStaticPaths: GetStaticPaths = async () => {
@@ -30,7 +28,7 @@ export const getStaticPaths: GetStaticPaths = async () => {
 }
 
 export const getStaticProps: GetStaticProps = async ({ params }) => {
-  const allPosts = getAllPosts([
+  const allPosts = getPosts([
     'title',
     'date',
     'slug',
